Added a new chart type "Category"
Adds a new type "category" which is a line graph, but with labels for the x axis rather than scaled numerical values.
This is to resolve #121, but still has a few issues.

Originally I copied PlotContentLineChart into a new PlotContentCategoryChart, but instead found that with just a few extra conditionals could allow for a large amount of reuse code.

There is still an issue where the angle points of the line graph still don't line up right with the X axis lines.

The "category" type does have a few restrictions:
  * Setting a min and max X axis value has no effect
  * Setting the X axis to be log has no effect
  * When providing data series the X data points must always match, missing data points for a given series is not currently handled