BugFix - NPE when plotting small numbers
Before this patch XChart would get an NPE when trying to plot numbers
which are smaller then 1E-10. This happened because gridStep was scaled
using a hard coded 10, zeroing out numbers on smaller scale.
Now we choose the right scale for small numbers and use that.

This fixes a crash when running the following code:
	double[] xData = new double[] { 1.0, 2.0};
	double[] yData = new double[] { 1E-12, 1.1E-12};
	Chart_XY chart = new ChartBuilder_XY().build();
	chart.addSeries("Test", xData, yData);
	new SwingWrapper(chart).displayChart();