Additional "Label" String to use in Legend
Proposed in #227, I added an extra String for Series. By default it is the same as "name" but can be changed with the method "setLabel".
Also I changed the return types of the Setter to Series so you can concatenate those (like it is in inheriting classes). So you can do it now like this: chart.addSeries(...).setLabel(...).setShowInLegend(true);...