Added a sumFormat attribute for PieStyler
In order to allow custom “Sum” formats within Pie graphs added a `sumFormat` attribute to the `PieStyler` if non-null this acts as a MessageFormat to calculate the string to be displayed as the sum.
If null (default) then the current behaviour will hold and the sum will be calculated as a number using the PieStyler DecimalFormat.
Within the MessageFormat string `{0}` acts as a placeholder for the sum value (number using the PieStyler DecimalFormat).

So for example if the sum for a Pie is 101 then calling.

`PieStyler.setSumFormat("Total {0} GB")`

will cause the sum to be displayed as `Total 101 GB`