Improve numerical stability of log axis tick calculation
For values below ~1e-8 on a log axis, a large number of ticks are
generated due to the fixed epsilon used in AxisTickCalculator_Logarithmic.
This commit simplifies the tick calculation to use integer steps between
each decade, preventing the need for inaccurate floating-point
comparisons as well as accumulation of FP error.